backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
email-validator==2.1.0.post1

# Database
motor==3.3.2
beanie==1.25.0
pymongo==4.6.1

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dateutil==2.8.2

# AI Integration
openai==1.12.0

# Development
pytest==8.0.1
pytest-asyncio==0.23.5
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Space API Dependencies
Pillow==10.4.0
opencv-python-headless==4.10.0.84
PyPDF2==3.0.1
python-magic==0.4.27
